+++ packages/azure-functions/ProcessTrustedFile/index.mjs
@@ -1,10 +1,9 @@
-import buildUploadConfigFromMessage from '../Shared/build-upload-config-from-message.js'
-import buildMetricConfigFromMessage from '../Shared/build-metric-config-from-message.js'
+import buildConfigFromMessage from '../Shared/build-config-from-message.js'
 import buildSignalRMessage from '../Shared/build-signalr-message.js'
 
 export default async function (context, message) {
   context.log('Processing', JSON.stringify(message))
-  const config = message.isMetricDataExtraction ? buildMetricConfigFromMessage(message) : buildUploadConfigFromMessage(message)
+  const config = buildConfigFromMessage(message)
   let processingFunction
   try {
     // Load the processing function for the upload type.
